build: re-organize directory structure
