+++ packages/semi-rspack/src/index.ts
@@ -1,5 +1,5 @@
 export * from './types';
-export * from './semi-webpack-plugin';
+export * from './plugin';
 export * from './rule';
 export { default as semiThemeLoader } from './loaders/semi-theme-loader';
 export { default as semiPrefixLoader } from './loaders/semi-prefix-loader';
